Добавил отдельным классом CustomDialog, реализовал показ диалогов перед удалением продуктов

+++ app/src/main/java/com/bryukhanov/shoppinglist/productslist/presentation/view/ProductsListFragment.kt
@@ -18,6 +18,7 @@ import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import com.bryukhanov.shoppinglist.R
 import com.bryukhanov.shoppinglist.core.util.Animates
+import com.bryukhanov.shoppinglist.core.util.CustomDialog
 import com.bryukhanov.shoppinglist.core.util.Units
 import com.bryukhanov.shoppinglist.databinding.FragmentProductsListBinding
 import com.bryukhanov.shoppinglist.productslist.domain.models.ProductListItem
@@ -214,11 +215,33 @@ class ProductsListFragment : Fragment() {
         }
 
         binding.deleteAllProductsMenu.setOnClickListener {
-            viewModel.deleteAllProduct(shoppingListId)
+            bottomSheetMenu?.state = BottomSheetBehavior.STATE_HIDDEN
+            if (productsAdapter.itemCount > 0) {
+                CustomDialog(requireContext()).showConfirmDialog(
+                    theme = R.style.CustomDialogTheme,
+                    message = getString(R.string.dialog_message_delete_all_product),
+                    positiveButtonText = getString(R.string.dialog_positive_answer),
+                    negativeButtonText = getString(R.string.dialog_cancel),
+                    onPositiveClick = {
+                        viewModel.deleteAllProduct(shoppingListId)
+                    }
+                )
+            }
         }
 
         binding.clearBoughtMenu.setOnClickListener {
-            viewModel.deleteBoughtProduct(shoppingListId)
+            bottomSheetMenu?.state = BottomSheetBehavior.STATE_HIDDEN
+            if (productsAdapter.itemCount > 0) {
+                CustomDialog(requireContext()).showConfirmDialog(
+                    theme = R.style.CustomDialogTheme,
+                    message = getString(R.string.dialog_message_delete_bought_product),
+                    positiveButtonText = getString(R.string.dialog_positive_answer),
+                    negativeButtonText = getString(R.string.dialog_cancel),
+                    onPositiveClick = {
+                        viewModel.deleteBoughtProduct(shoppingListId)
+                    }
+                )
+            }
         }
 
         binding.overlay.setOnClickListener {
